# backend/app/routes/admin/settings.py
from datetime import datetime
from typing import Annotated, Any, cast

from fastapi import APIRouter, Depends, HTTPException, Query, Request

from app.dependencies import get_async_supabase_admin_client
from app.logger import logger, sanitize_log_value
from app.middleware.auth_middleware import require_permission
from app.routes.admin.helpers import ADMIN_ERROR_RESPONSES
from app.schemas.settings_schemas import (
    ConfigHistoryResponse,
    ConfigResponse,
    ConfigUpdate,
    PendingActionRequest,
    PendingConfigChangeResponse,
)
from app.schemas.user import User
from app.services.email_service import email_service
from app.services.encryption_service import encryption_service
from app.services.line_service import line_service
from app.services.redis_service import redis_service

router = APIRouter()


# Versioned so a deployment cannot serve a pending-settings cache populated by
# the pre-redaction response shape.
PENDING_SETTINGS_CACHE_KEY = "admin_settings_pending:v2"


def _settings_cache_key(category: str | None) -> str:
    return f"admin_settings:{category or 'all'}"


def _settings_history_cache_key(key: str) -> str:
    return f"admin_settings_history:{key}"


async def _invalidate_settings_cache(config_key: str | None = None) -> None:
    await redis_service.delete_pattern("admin_settings:*")
    await redis_service.delete(PENDING_SETTINGS_CACHE_KEY)
    if config_key:
        await redis_service.delete(_settings_history_cache_key(config_key))


SYSTEM_CONFIG_SELECT_FIELDS = (
    "key, value, type, description, category, is_public, is_encrypted, requires_approval, updated_at, updated_by"
)


def _mask_encrypted_value(value: Any, is_encrypted: bool) -> Any:
    return None if is_encrypted else value


async def _record_config_history(
    admin_client: Any,
    config_key: str,
    old_value: Any,
    new_value: Any,
    changed_by: str,
    reason: str,
) -> None:
    await (
        admin_client.table("config_history")
        .insert(
            {
                "config_key": config_key,
                "old_value": old_value,
                "new_value": new_value,
                "changed_by": changed_by,
                "change_reason": reason,
            }
        )
        .execute()
    )


def _prepare_setting_value(key: str, setting: dict[str, Any], update_data: ConfigUpdate) -> tuple[Any, Any, bool]:
    value_to_store = update_data.value
    old_value = setting["value"]
    is_encrypted = bool(setting.get("is_encrypted", False))
    if not is_encrypted:
        return value_to_store, old_value, False
    if update_data.value is None or (isinstance(update_data.value, str) and not update_data.value.strip()):
        raise HTTPException(status_code=400, detail="Encrypted setting replacement cannot be empty")
    try:
        value_to_store = encryption_service.encrypt_value(update_data.value, setting.get("type", "string"))
    except Exception as exc:
        logger.error("Failed to encrypt setting %s: %s", sanitize_log_value(key), sanitize_log_value(str(exc)))
        raise HTTPException(status_code=500, detail="Failed to encrypt sensitive setting") from exc
    return value_to_store, old_value, True


async def _create_pending_setting_change(
    admin_client: Any,
    key: str,
    value_to_store: Any,
    current_admin: User,
    is_encrypted: bool,
) -> dict[str, Any]:
    pending = await (
        admin_client.table("pending_config_changes")
        .insert(
            {
                "config_key": key,
                "proposed_value": value_to_store,
                "requester_id": current_admin.id,
                "status": "pending",
            }
        )
        .execute()
    )
    requester_name = current_admin.name or current_admin.email
    await line_service.send_notification(
        f"\n[PURRFECT ADMIN]\n⚠️ Approval Required\nSetting: {key}\nRequested by: {requester_name}"
    )
    await _invalidate_settings_cache(key)
    pending_response = dict(cast(dict[str, Any], pending.data[0]))
    if is_encrypted:
        pending_response["proposed_value"] = None
    return pending_response


@router.get("", response_model=list[ConfigResponse], responses=ADMIN_ERROR_RESPONSES)
async def get_all_settings(
    current_admin: Annotated[User, Depends(require_permission("system:settings"))],
    category: Annotated[str | None, Query()] = None,
    cache_bust: Annotated[str | None, Query()] = None,
) -> list[dict[str, Any]]:
    """Get all system settings with metadata."""
    try:
        cache_key = _settings_cache_key(category)
        if not cache_bust:
            cached = await redis_service.get(cache_key)
        else:
            cached = None
        if cached is not None:
            raw_items = cached
        else:
            admin_client = await get_async_supabase_admin_client()
            query = (
                admin_client.table("system_configs").select(SYSTEM_CONFIG_SELECT_FIELDS).order("category").order("key")
            )
            if category:
                query = query.eq("category", category)
            result = await query.execute()
            raw_items = result.data or []
            if not cache_bust:
                await redis_service.set(cache_key, raw_items, expire=120)

        # SECURITY: Keep encrypted values server-side. Admins can replace a
        # secret without receiving its plaintext through the browser.
        masked_data = []
        for cached_item in raw_items:
            item = dict(cached_item)
            if item.get("is_encrypted"):
                item["value"] = None
            masked_data.append(item)

        return cast(list[dict[str, Any]], masked_data)
    except Exception as e:
        logger.error("Failed to fetch settings: %s", e)
        raise HTTPException(status_code=500, detail="Failed to fetch system settings")


@router.get("/history/{key}", response_model=list[ConfigHistoryResponse], responses=ADMIN_ERROR_RESPONSES)
async def get_setting_history(
    key: str, current_admin: Annotated[User, Depends(require_permission("system:settings"))]
) -> list[dict[str, Any]]:
    """Get evolution history for a specific setting."""
    try:
        admin_client = await get_async_supabase_admin_client()
        config_result = (
            await admin_client.table("system_configs").select("is_encrypted").eq("key", key).maybe_single().execute()
        )
        config_data = cast(dict[str, Any], config_result.data) if config_result and config_result.data else {}
        # If metadata is missing, do not assume a historical value is safe to
        # disclose. Masking is the fail-closed response for stale records.
        is_encrypted = bool(config_data.get("is_encrypted")) if config_data else True

        cache_key = _settings_history_cache_key(key)
        cached = await redis_service.get(cache_key)
        if cached is not None:
            cached_entries = []
            for cached_item in cast(list[dict[str, Any]], cached):
                entry = dict(cached_item)
                entry["old_value"] = _mask_encrypted_value(entry.get("old_value"), is_encrypted)
                entry["new_value"] = _mask_encrypted_value(entry.get("new_value"), is_encrypted)
                cached_entries.append(entry)
            return cached_entries

        result = (
            await admin_client.table("config_history")
            .select(
                "id, config_key, old_value, new_value, changed_by, change_reason, created_at, user:users!changed_by(email)"
            )
            .eq("config_key", key)
            .order("created_at", desc=True)
            .execute()
        )
        history_entries = []
        for item in result.data or []:
            entry = cast(dict[str, Any], item)
            user_info = entry.pop("user", None)
            entry["user_email"] = user_info.get("email") if isinstance(user_info, dict) else None
            entry["old_value"] = _mask_encrypted_value(entry.get("old_value"), is_encrypted)
            entry["new_value"] = _mask_encrypted_value(entry.get("new_value"), is_encrypted)
            history_entries.append(entry)
        await redis_service.set(cache_key, history_entries, expire=120)
        return history_entries
    except Exception as e:
        logger.error("Failed to fetch history for %s: %s", sanitize_log_value(key), sanitize_log_value(str(e)))
        raise HTTPException(status_code=500, detail="Failed to fetch config history")


@router.put("/{key}", response_model=ConfigResponse | PendingConfigChangeResponse, responses=ADMIN_ERROR_RESPONSES)
async def update_setting(
    request: Request,
    key: str,
    update_data: ConfigUpdate,
    current_admin: Annotated[User, Depends(require_permission("system:settings"))],
) -> dict[str, Any]:
    """Update a setting OR create an approval request if required."""
    try:
        admin_client = await get_async_supabase_admin_client()

        check = (
            await admin_client.table("system_configs")
            .select("value, requires_approval, type, is_encrypted")
            .eq("key", key)
            .single()
            .execute()
        )
        if not check.data:
            raise HTTPException(status_code=404, detail="Setting not found")

        setting = cast(dict[str, Any], check.data)

        value_to_store, old_value_for_history, is_secret = _prepare_setting_value(key, setting, update_data)

        # Check if approval is required (Maker-Checker)
        if setting.get("requires_approval", False):
            return await _create_pending_setting_change(admin_client, key, value_to_store, current_admin, is_secret)

        # Immediate update if no approval needed
        update_values = {"value": value_to_store, "updated_by": current_admin.id}
        if update_data.description is not None:
            update_values["description"] = update_data.description
        if update_data.category is not None:
            update_values["category"] = update_data.category

        result = (
            await cast(Any, admin_client.table("system_configs").update(update_values))
            .eq("key", key)
            .select(SYSTEM_CONFIG_SELECT_FIELDS)
            .single()
            .execute()
        )

        # Log to Detailed Config History (sanitize encrypted secrets)
        history_old = "[ENCRYPTED_SECRET]" if is_secret else old_value_for_history
        history_new = "[ENCRYPTED_SECRET]" if is_secret else value_to_store

        await _record_config_history(
            admin_client,
            config_key=key,
            old_value=history_old,
            new_value=history_new,
            changed_by=current_admin.id,
            reason="Direct administrative update",
        )
        await _invalidate_settings_cache(key)
        response_data = dict(cast(dict[str, Any], result.data))
        response_data["value"] = _mask_encrypted_value(response_data.get("value"), is_secret)
        return response_data
    except HTTPException:
        raise
    except Exception as e:
        logger.error("Failed to update setting %s: %s", sanitize_log_value(key), sanitize_log_value(str(e)))
        raise HTTPException(status_code=500, detail="Failed to update system setting")


@router.get("/pending", response_model=list[PendingConfigChangeResponse], responses=ADMIN_ERROR_RESPONSES)
async def get_pending_changes(
    current_admin: Annotated[User, Depends(require_permission("system:settings"))],
    cache_bust: Annotated[str | None, Query()] = None,
) -> list[dict[str, Any]]:
    """Get all pending config changes (Checkers)"""
    try:
        cached = None if cache_bust else await redis_service.get(PENDING_SETTINGS_CACHE_KEY)
        if cached is not None:
            return cast(list[dict[str, Any]], cached)

        admin_client = await get_async_supabase_admin_client()
        result = (
            await admin_client.table("pending_config_changes")
            .select(
                "id, config_key, proposed_value, requester_id, approver_id, status, rejection_reason, created_at, updated_at, requester:users!requester_id(email)"
            )
            .eq("status", "pending")
            .execute()
        )

        pending_changes = cast(list[dict[str, Any]], result.data or [])
        config_keys = list({str(item["config_key"]) for item in pending_changes})
        current_values_by_key: dict[str, object | None] = {}
        encrypted_keys: set[str] = set()
        known_config_keys: set[str] = set()

        if config_keys:
            configs_res = await (
                admin_client.table("system_configs")
                .select("key, value, is_encrypted")
                .in_("key", config_keys)
                .execute()
            )
            configs_res_data = cast(list[dict[str, Any]], configs_res.data or [])
            known_config_keys = {str(config["key"]) for config in configs_res_data}
            current_values_by_key = {
                str(config["key"]): _mask_encrypted_value(config.get("value"), bool(config.get("is_encrypted")))
                for config in configs_res_data
            }
            encrypted_keys = {str(config["key"]) for config in configs_res_data if bool(config.get("is_encrypted"))}

        normalized_changes = []
        for item in pending_changes:
            requester_info = item.pop("requester", None)
            item["requester_email"] = requester_info.get("email") if isinstance(requester_info, dict) else None
            item["current_value"] = current_values_by_key.get(item["config_key"])
            if item["config_key"] in encrypted_keys or item["config_key"] not in known_config_keys:
                item["proposed_value"] = None
            normalized_changes.append(item)

        if not cache_bust:
            await redis_service.set(PENDING_SETTINGS_CACHE_KEY, normalized_changes, expire=60)
        return normalized_changes
    except Exception as e:
        logger.error("Failed to fetch pending changes: %s", e)
        raise HTTPException(status_code=500, detail="Failed to fetch pending changes")


async def _fetch_pending_change(admin_client: Any, change_id: str) -> dict[str, Any]:
    """Fetch pending config change or raise 404."""
    pending_check = (
        await admin_client.table("pending_config_changes").select("*").eq("id", change_id).single().execute()
    )
    pending_data = cast(dict[str, Any], pending_check.data) if pending_check.data else None
    if not pending_data or pending_data.get("status") != "pending":
        raise HTTPException(status_code=404, detail="Pending change not found or already processed")
    return pending_data


async def _notify_change_requester(
    admin_client: Any, change: dict[str, Any], status: str, admin: User, reason: str | None = None
) -> None:
    """Notify requester of config change result."""
    requester_res = (
        await admin_client.table("users").select("email").eq("id", change["requester_id"]).single().execute()
    )
    req_data = cast(dict[str, Any], requester_res.data) if requester_res.data else None
    if req_data:
        if reason:
            email_service.send_admin_config_result(
                str(req_data.get("email")),
                str(change.get("config_key")),
                status,
                str(admin.name or admin.email),
                reason,
            )
        else:
            email_service.send_admin_config_result(
                str(req_data.get("email")),
                str(change.get("config_key")),
                status,
                str(admin.name or admin.email),
            )


@router.post("/approve/{change_id}", response_model=ConfigResponse, responses=ADMIN_ERROR_RESPONSES)
async def approve_change(
    change_id: str,
    current_admin: Annotated[User, Depends(require_permission("system:settings"))],
    payload: PendingActionRequest | None = None,
) -> dict[str, Any]:
    """Approve a pending config change (Checker)."""
    try:
        admin_client = await get_async_supabase_admin_client()

        # 1. Fetch pending change
        change = await _fetch_pending_change(admin_client, change_id)

        # Prevent self-approval (Maker cannot be Checker)
        if str(change["requester_id"]) == str(current_admin.id):
            raise HTTPException(status_code=403, detail="You cannot approve your own changes")

        # 2. Fetch current config
        config_check = (
            await admin_client.table("system_configs").select("*").eq("key", change["config_key"]).single().execute()
        )
        current_config = cast(dict[str, Any], config_check.data)

        # 3. Update main config
        update_result = (
            await cast(
                Any,
                admin_client.table("system_configs").update(
                    {"value": change["proposed_value"], "updated_by": current_admin.id}
                ),
            )
            .eq("key", change["config_key"])
            .select(SYSTEM_CONFIG_SELECT_FIELDS)
            .single()
            .execute()
        )

        # 4. Record History
        is_secret = bool(current_config.get("is_encrypted"))
        await _record_config_history(
            admin_client,
            config_key=change["config_key"],
            old_value=_mask_encrypted_value(current_config.get("value"), is_secret),
            new_value=_mask_encrypted_value(change.get("proposed_value"), is_secret),
            changed_by=current_admin.id,
            reason=f"Approved from Request {change_id}",
        )

        # 5. Update Pending record
        await (
            admin_client.table("pending_config_changes")
            .update({"status": "approved", "approver_id": current_admin.id, "updated_at": datetime.now().isoformat()})
            .eq("id", change_id)
            .execute()
        )

        # 6. Notify Requester
        await _notify_change_requester(admin_client, change, "approved", current_admin)

        await _invalidate_settings_cache(change["config_key"])
        response_data = dict(cast(dict[str, Any], update_result.data))
        response_data["value"] = _mask_encrypted_value(response_data.get("value"), is_secret)
        return response_data
    except Exception as e:
        if isinstance(e, HTTPException):
            raise e
        logger.error("Approval failed for change %s: %s", sanitize_log_value(change_id), sanitize_log_value(str(e)))
        raise HTTPException(status_code=500, detail="Approval process failed")


@router.post("/reject/{change_id}", response_model=dict, responses=ADMIN_ERROR_RESPONSES)
async def reject_change(
    change_id: str,
    current_admin: Annotated[User, Depends(require_permission("system:settings"))],
    payload: PendingActionRequest,
) -> dict[str, Any]:
    """Reject a pending config change."""
    try:
        admin_client = await get_async_supabase_admin_client()

        # 1. Fetch pending change first to get requester info
        change = await _fetch_pending_change(admin_client, change_id)

        # 2. Update status to rejected
        await (
            admin_client.table("pending_config_changes")
            .update(
                {
                    "status": "rejected",
                    "approver_id": current_admin.id,
                    "rejection_reason": payload.rejection_reason,
                    "updated_at": datetime.now().isoformat(),
                }
            )
            .eq("id", change_id)
            .execute()
        )

        # 3. Notify Requester
        await _notify_change_requester(
            admin_client, change, "rejected", current_admin, str(payload.rejection_reason or "")
        )

        await _invalidate_settings_cache(change["config_key"])
        return {"status": "rejected", "change_id": change_id}
    except HTTPException:
        raise
    except Exception as e:
        logger.error("Rejection failed for %s: %s", sanitize_log_value(change_id), sanitize_log_value(str(e)))
        raise HTTPException(status_code=500, detail="Rejection process failed")