# docs/openapi-baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PurrFect Spots API",
    "description": "\n        PurrFect Spots API helps you share and discover cat locations in a socially-connected way.\n\n        ### 🌐 API Specification\n        All endpoints follow the `/api/v1/` prefix convention for stable identification.\n\n        ### ✨ Key Features\n        * 📍 **Geo-Location Awareness**: High-precision cat spot sharing with privacy protection.\n        * 🤳 **AI Photo Intelligence**: Automated cat detection and visual attribute extraction.\n        * 🔒 **Enterprise-Grade Security**: PKCE-enabled OAuth 2.0 and Supabase-backed authentication.\n        * ⚡ **Performance Optimized**: Built-in rate limiting, GZip compression, and ETag support.\n        * 🔄 **Real-time Ready**: Designed for low-latency interactions and live updates.\n        ",
    "contact": {
      "name": "Purrfect Spots Team",
      "email": "support@purrfectspots.com"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "3.1.0"
  },
  "paths": {
...
    "/api/v1/admin/settings": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get All Settings",
        "description": "Get all system settings with metadata.",
        "operationId": "get_all_settings_api_v1_admin_settings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "cache_bust",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cache Bust"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConfigResponse"
                  },
                  "title": "Response Get All Settings Api V1 Admin Settings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/settings/history/{key}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Setting History",
        "description": "Get evolution history for a specific setting.",
        "operationId": "get_setting_history_api_v1_admin_settings_history__key__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConfigHistoryResponse"
                  },
                  "title": "Response Get Setting History Api V1 Admin Settings History  Key  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/settings/{key}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update Setting",
        "description": "Update a setting OR create an approval request if required.",
        "operationId": "update_setting_api_v1_admin_settings__key__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConfigResponse"
                    },
                    {
                      "$ref": "#/components/schemas/PendingConfigChangeResponse"
                    }
                  ],
                  "title": "Response Update Setting Api V1 Admin Settings  Key  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/settings/pending": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Pending Changes",
        "description": "Get all pending config changes (Checkers)",
        "operationId": "get_pending_changes_api_v1_admin_settings_pending_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cache_bust",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cache Bust"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PendingConfigChangeResponse"
                  },
                  "title": "Response Get Pending Changes Api V1 Admin Settings Pending Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/settings/approve/{change_id}": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Approve Change",
        "description": "Approve a pending config change (Checker).",
        "operationId": "approve_change_api_v1_admin_settings_approve__change_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "change_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Change Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PendingActionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/settings/reject/{change_id}": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Reject Change",
        "description": "Reject a pending config change.",
        "operationId": "reject_change_api_v1_admin_settings_reject__change_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "change_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Change Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PendingActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reject Change Api V1 Admin Settings Reject  Change Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigHistoryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "config_key": {
            "type": "string",
            "title": "Config Key"
          },
          "old_value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Old Value"
          },
          "new_value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "New Value"
          },
          "changed_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed By"
          },
          "user_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Email"
          },
          "change_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "config_key",
          "created_at"
        ],
        "title": "ConfigHistoryResponse"
      },
      "ConfigResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "title": "Value"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "default": false
          },
          "is_encrypted": {
            "type": "boolean",
            "title": "Is Encrypted",
            "default": false
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": false
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "ConfigResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PendingActionRequest": {
        "properties": {
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          }
        },
        "type": "object",
        "title": "PendingActionRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}